requirements.txt
flask==2.3.*
python-dotenv==1.*
beautifulsoup4==4.*
validators==0.*
psycopg2-binary==2.*
requests==2.*